# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LibraryOS API - Multi-tenant school bookstore management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for current tenant
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: showInStore
          in: query
          schema:
            type: boolean
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}/toggle-store:
    post:
      operationId: toggleProductStore
      tags: [products]
      summary: Toggle product visibility in online store
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visibility toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Find product by barcode (for POS scanning)
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========== CARDS ==========
...
  /platforms:
    get:
      operationId: listPlatforms
      tags: [platforms]
      summary: List all educational platforms
      responses:
        "200":
          description: List of platforms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"
    post:
      operationId: createPlatform
      tags: [platforms]
      summary: Create educational platform (Super Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlatformRequest"
      responses:
        "201":
          description: Platform created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
  /platforms/{platformId}:
    patch:
      operationId: updatePlatform
      tags: [platforms]
      summary: Update educational platform
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlatformRequest"
      responses:
        "200":
          description: Platform updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    delete:
      operationId: deletePlatform
      tags: [platforms]
      summary: Delete educational platform
      parameters:
        - name: platformId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ========== SHIFTS ==========
...
  /print-services:
    get:
      operationId: listPrintServices
      tags: [print-services]
      summary: List print services for current tenant
      responses:
        "200":
          description: List of print services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrintService"
    post:
      operationId: createPrintService
      tags: [print-services]
      summary: Create a print service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrintServiceRequest"
      responses:
        "201":
          description: Print service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintService"
  /print-services/{serviceId}:
    patch:
      operationId: updatePrintService
      tags: [print-services]
      summary: Update a print service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrintServiceRequest"
      responses:
        "200":
          description: Print service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintService"
    delete:
      operationId: deletePrintService
      tags: [print-services]
      summary: Delete a print service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ========== ORDERS (Online Store) ==========
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        category:
          type: string
        barcode:
          type: ["string", "null"]
        price:
          type: number
        stockQuantity:
          type: integer
        stockAlertThreshold:
          type: integer
        imageUrl:
          type: ["string", "null"]
        showInStore:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - category
        - price
        - stockQuantity
        - stockAlertThreshold
        - showInStore
        - isActive
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        barcode:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        stockAlertThreshold:
          type: integer
        imageUrl:
          type: string
        showInStore:
          type: boolean
      required:
        - name
        - category
        - price
        - stockQuantity
        - stockAlertThreshold
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        barcode:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        stockAlertThreshold:
          type: integer
        imageUrl:
          type: string
        showInStore:
          type: boolean
        isActive:
          type: boolean
...
    Platform:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        grades:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        pricingTiers:
          type: array
          items:
            $ref: "#/components/schemas/PlatformPricingTier"
        currentOffer:
          type: ["string", "null"]
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - grades
        - subjects
        - pricingTiers
        - isActive
        - updatedAt
    PlatformPricingTier:
      type: object
      properties:
        cardValue:
          type: number
        price:
          type: number
        discountedPrice:
          type: ["number", "null"]
      required:
        - cardValue
        - price
    CreatePlatformRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        grades:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        pricingTiers:
          type: array
          items:
            $ref: "#/components/schemas/PlatformPricingTier"
        currentOffer:
          type: string
      required:
        - name
        - grades
        - subjects
        - pricingTiers
    UpdatePlatformRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        grades:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        pricingTiers:
          type: array
          items:
            $ref: "#/components/schemas/PlatformPricingTier"
        currentOffer:
          type: string
        isActive:
          type: boolean

    # SHIFT
...
    PrintService:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        pricingType:
          type: string
          enum: [fixed, per_page]
        price:
          type: number
        paperSize:
          type: ["string", "null"]
        colorType:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - pricingType
        - price
        - isActive
        - createdAt
    CreatePrintServiceRequest:
      type: object
      properties:
        name:
          type: string
        pricingType:
          type: string
          enum: [fixed, per_page]
        price:
          type: number
        paperSize:
          type: string
        colorType:
          type: string
      required:
        - name
        - pricingType
        - price
    UpdatePrintServiceRequest:
      type: object
      properties:
        name:
          type: string
        pricingType:
          type: string
          enum: [fixed, per_page]
        price:
          type: number
        paperSize:
          type: string
        colorType:
          type: string
        isActive:
          type: boolean

    # ORDER
...